refactor: getBoardDetail 병합
