+++ apps/community/src/constants/api.ts
@@ -8,7 +8,7 @@ const MOCK_END_POINT = {
   CONTACT: `${MOCK_BASE_URL}/about/contact`,
   MY_PROFILE: `${MOCK_BASE_URL}/my`,
   BOARD: `${MOCK_BASE_URL}/board`,
-  NOTICE_DETAIL: (id: string) => `${MOCK_BASE_URL}/posts/${id}`,
+  BOARD_DETAIL: (id: string) => `${MOCK_BASE_URL}/posts/${id}`,
 } as const;
 
 export { MOCK_BASE_URL, MOCK_END_POINT };
